Lag ny fil "UtbetalingsperiodeUtil" og flytt hjelpefunskjoner for utbetalingsperioder dit
